# Conflicts:
#	han_back/api/__pycache__/__init__.cpython-37.pyc
#	han_back/api/__pycache__/admin.cpython-37.pyc
#	han_back/api/__pycache__/urls.cpython-37.pyc
#	han_back/api/adminModels/__pycache__/commentAdmin.cpython-37.pyc
#	han_back/api/migrations/__pycache__/__init__.cpython-37.pyc
#	han_back/api/models/__pycache__/__init__.cpython-37.pyc
#	han_back/api/models/__pycache__/articleLikeModel.cpython-37.pyc
#	han_back/api/models/__pycache__/comment.cpython-37.pyc
#	han_back/api/serializers/__pycache__/__init__.cpython-37.pyc
#	han_back/api/serializers/__pycache__/articleLikeSerializer.cpython-37.pyc
#	han_back/api/serializers/__pycache__/categorySerializer.cpython-37.pyc
#	han_back/api/serializers/__pycache__/commentSerializer.cpython-37.pyc
#	han_back/api/serializers/__pycache__/userSerializer.cpython-37.pyc
#	han_back/api/views/__pycache__/articleLikeViews.cpython-37.pyc
#	han_back/api/views/__pycache__/articleView.cpython-37.pyc
#	han_back/api/views/__pycache__/auth.cpython-37.pyc
#	han_back/api/views/__pycache__/welcomeView.cpython-37.pyc
#	han_back/db.sqlite3
#	han_back/han_back/__pycache__/__init__.cpython-37.pyc
#	han_back/han_back/__pycache__/settings.cpython-37.pyc
#	han_back/han_back/__pycache__/urls.cpython-37.pyc
#	han_back/han_back/__pycache__/wsgi.cpython-37.pyc

+++ han_back/api/views/commentView.py
@@ -44,4 +44,7 @@ class CommentDelete(generics.DestroyAPIView):
         return Comment.objects.get(id=self.kwargs[self.lookup_field])
 
     def perform_destroy(self, instance):
-        instance.delete()
+        if self.request.user == self.get_object().created_by:
+            instance.delete()
+        else:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
